+++ MEAN/source_angular/bookMyBusApp/src/app/adminComponent/adminComp.ts
@@ -2,6 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
 import { Bus } from './businterface';
 import { BusService } from './bus_service';
+import { UserService } from '../userComponent/user_service';
 
 
 @Component({
@@ -18,8 +19,10 @@ export class AdminComp implements OnInit{
     tocityv: boolean = true;
     capacityv: boolean = true;
     buses: Bus[];
+    resp: any = {};
+    routeno: any;
 
-    constructor(private _busService: BusService, private router: Router){
+    constructor(private _busService: BusService, private _userService: UserService, private router: Router){
         this.getBuses();           // for reloading page when navigated here from other component
     }
 
@@ -38,17 +41,51 @@ export class AdminComp implements OnInit{
     updatebus(id){
 
         // check if bus is not booked for present/future dates, then update else not
-
-        this.router.navigate(["adminconsole/buses/editbus",id]);
+        this._busService.getBus(id).subscribe(
+            (resp: any)=>{
+                this.resp = resp;
+                this.routeno = this.resp.routeno;
+                this._userService.checkBusForUD(this.routeno).subscribe(
+                    (resp: any) =>{
+                        this.resp = resp;
+                        if(this.resp.counter == 0){
+                            this.router.navigate(["adminconsole/buses/editbus",id]);
+                        }
+                        else{
+                            alert("CANNOT UPDATE BUS as it is BOOKED.");
+                        }
+                    },
+                    err => console.log(err)
+                );
+            },
+            err => console.log(err)
+        );
     }
 
     deletebus(id){
 
         // check if bus is not booked for present/future dates, then delete else not
-
-        this._busService.deleteBus(id).subscribe(
-            (data: any) => {
-                this.getBuses();
+        this._busService.getBus(id).subscribe(
+            (resp: any)=>{
+                this.resp = resp;
+                this.routeno = this.resp.routeno;
+                this._userService.checkBusForUD(this.routeno).subscribe(
+                    (resp: any) =>{
+                        this.resp = resp;
+                        if(this.resp.counter == 0){
+                            this._busService.deleteBus(id).subscribe(
+                                (data: any) => {
+                                    this.getBuses();
+                                },
+                                err => console.log(err)
+                            );
+                        }
+                        else{
+                            alert("CANNOT DELETE BUS as it is BOOKED.");
+                        }
+                    },
+                    err => console.log(err)
+                );
             },
             err => console.log(err)
         );
